# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Banco Demo API",
    "version": "1.0.0",
    "description": "Backend bancário demo (contrato-compartilhado v1.0). Sessão por cookie bank_session. Transferências assíncronas via Saga: POST → PENDING/PROCESSING → GET → COMPLETED/FAILED."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3001",
      "description": "Backend local"
    }
  ],
  "paths": {
...
    "/v1/transfers": {
      "post": {
        "summary": "Solicitar transferência (202 novo, 200 replay)",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Obrigatório. 8–128 chars [A-Za-z0-9._:-]. Escopo por conta remetente, sem expiração. Replay com mesmo payload → 200 + mesma operação; payload diferente → 409.",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._:-]{8,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestTransferBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replay idempotente (mesmo ID, estado atual)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            }
          },
          "202": {
            "description": "Aceita (PENDING)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RECIPIENT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "IDEMPOTENCY_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "SELF_TRANSFER",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Histórico paginado (só enviadas)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Inteiro 1–50, default 20",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{1,3}$",
              "default": "20"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Cursor opaco da página anterior",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferPage"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR (limit/cursor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transfers/{id}": {
      "get": {
        "summary": "Detalhe da transferência (só enviadas)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "TRANSFER_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bankSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bank_session",
        "description": "Sessão opaca HttpOnly, 24h. Signup/signin emitem Set-Cookie; signout expira."
      }
    },
    "schemas": {
...
      "Transfer": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "sourceAccountId",
          "recipientAccountId",
          "recipientName",
          "amountCents",
          "currency",
          "note",
          "status",
          "failureCode",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "sourceAccountId": {
            "type": "string"
          },
          "recipientAccountId": {
            "type": "string"
          },
          "recipientName": {
            "type": "string"
          },
          "amountCents": {
            "type": "integer"
          },
          "currency": {
            "type": "string",
            "const": "BRL"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "failureCode": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "INSUFFICIENT_FUNDS",
              "CREDIT_FAILED",
              null
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "TransferPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "nextCursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "id",
                "sourceAccountId",
                "recipientAccountId",
                "recipientName",
                "amountCents",
                "currency",
                "note",
                "status",
                "failureCode",
                "createdAt",
                "updatedAt"
              ],
              "properties": {
                "id": {
                  "type": "string"
                },
                "sourceAccountId": {
                  "type": "string"
                },
                "recipientAccountId": {
                  "type": "string"
                },
                "recipientName": {
                  "type": "string"
                },
                "amountCents": {
                  "type": "integer"
                },
                "currency": {
                  "type": "string",
                  "const": "BRL"
                },
                "note": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "PENDING",
                    "PROCESSING",
                    "COMPLETED",
                    "FAILED"
                  ]
                },
                "failureCode": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "enum": [
                    "INSUFFICIENT_FUNDS",
                    "CREDIT_FAILED",
                    null
                  ]
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                }
              }
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "field",
                    "message"
                  ],
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "RequestTransferBody": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "recipientAccountId",
          "amountCents"
        ],
        "properties": {
          "recipientAccountId": {
            "type": "string",
            "minLength": 1
          },
          "amountCents": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000000
          },
          "note": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 140
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bankSession": []
    }
  ]
}